spaceModel.cpp
take user input save and return spacial characteristics to be used in spacial parameters.  This does not include physical or chemical composition requirements or partial pde modeling of the system.  this code set must be taken as a partial differential with respect to the second subset of characteristics held constant.  then, partial differential analysis of the other code set must be integrated while keeping this code set constant, over set range of integration.  final yield must be 3 dimensional volumetric model and physical characteristic ranges that describe the material itself.  Reah out to Joe for more information on how to use this in your 3d render generation programming